% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{select_cluster}
\alias{select_cluster}
\alias{select_cluster.data.frame}
\title{select_cluster}
\usage{
select_cluster(x, ...)

\method{select_cluster}{data.frame}(x, ...)
}
\arguments{
\item{x}{object to be passed to the shiny app.}

\item{...}{futher arguments passed down to methods.}
}
\description{
select_cluster
}
